Refactor backend file structure
