+++ src/backend/routes/authRoutes.ts
@@ -1,9 +1,14 @@
 import express from 'express';
 
-import { generateRandomString } from './utils';
-import { createSso, validateSsoRequest, createToken, generateUserDataFromSsoRequest } from './auth';
-import { upsertUser } from './models/users';
-import { getConfig } from './config';
+import { generateRandomString } from '../utils';
+import {
+  createSso,
+  validateSsoRequest,
+  createToken,
+  generateUserDataFromSsoRequest,
+} from '../auth';
+import { upsertUser } from '../models/users';
+import { getConfig } from '../config';
 
 const router = express.Router();
